[NF] Removing ignored library dirs for migration to git submodules
**NB:** This is a major version bump as we are changing the manner in
which third party libraries are included with IXP Manager.

Upgrade to this tag specifically first, then to the next tag (v3.1.1)
and follow the instructions there.

We use a large number of external / third party libraries and it's
very easy to get out of sync with these. By using Git submodules we
can make the install procedure easier as well as ensure the known
supported versions of third party libraries are used.

After upgrading to this version, remove the following directories from
your `library/` directory: