[NF] Add FANOUT type port for resellers
For resellers, we need to enforce the one port - one mac - one address rule
on the peering LAN. Depending on switch technology, this will be done using
a virtual ethernet port or a dedcaited fanout switch. A fanout port is a port
that sends a resold member's traffic to a peering port / switch.

I.e. it is an output port to the exchange to connects to a standard peering
input port.