[NF] Mailing List management via APIv1 (WIP)
A push for backup purposes. Work in progress of:

(a) removing the mailing list features from the CliController which I want to kill of (the generic controller that is, not the CLI access);
(b) abstraction of code so that the API can use the same code to manage mailing lists via wget / curl

This removes the requirement to SCP various files and commands around or to have IXP Manager installed / nfs'd on the mailman server.