Zend Framework 1 fallback
IXP Manager is transitioning from ZF1 to Laravel as a framework. It
will take some time to move over everything and this will be done on
an as needed basis. Realistically this means there may be some ZF1
crud for the foreseeable future.

We'll revert to ZF1 handling when Laravel throws a 404.