Make update-l2database.pl lookup lib and conf file
The Perl script update-l2database.pl, which is used to update mac
addresses learned by exchange point switches, is changed so that:
- It can be run from inside its original location in tools/runtime/
l2database. This is accomplished by the first 'use lib...' statement,
which tries to find the library to the relative directory ../../
perl-lib/IXPManager/lib. This means that the administrator may choose
not to install the library to the system, thus avoiding spreading
files in /usr/ locations.
- It also searches for libraries in ../lib/perl5, anticipating setups
where the script may be copied in a directory outside the IXP repo.
In that directory, the update-l2database.pl should be placed under
the bin/ subdirectory and the IXPManager library dir should be placed
under the lib/perl5/ subdirectory.
- The ixpmanager.conf file is also looked up in several locations.
Fist, in accordance to the previously mentioned scenario, the ../etc
relative directory is searched. Then, the /usr/local/etc and then the
/etc paths are searched. If none of the three contains the
ixpmanager.conf file, the script will complain and exit.