PPP - Split / Delete / Move Ports - closes islandbridgenetworks/IXP-Manager#36
So, we need to be able to do some super administrative things with ports.

Can we add a new option to the port drop down called 'Admin Actions' 

The additional options should be:

## Delete Port

This should yield a pop-up with:

> WARNING: Deletion is permanent and will remove the port from the patch panel including all history and files. [if duplex, add: As this is a duplex port, both individual ports will be deleted. If you do not want this, then split the port first.]

If the user proceeds, then delete the port(s) from history first and then the port itself.

## Split Port (only for duplex ports)

This option should remove the linked port from the master and reset it as available. Add a private note to the remaining / master port as follows:

> YYYY-MM-DD IXP Manager
> 
> This port was a duplex port with XXX and was split by YYY on YYYY-MM-DD.

Popup confirmation:

> Are you sure you want to split this port? The slave port (F4) will be removed from the master port (F3) and marked as available. If you want to split the other way (F4 as master), split now and then use the move function on F3 afterwards.


## Move Port

Sometimes a port fails / patch panel is removed / upgraded / etc.

We need to be able to move all the data (not historical, just active) of a patch panel port to another port of the same medium (utp / mmf / smf ) in the same location. Including if it's a duplex port.

For this we need to have an action that looks for the new patch panel and port(s).