[NF] Remove the meeting controller and all associated files
The meeting controller was written in house at INEX for our own particular
purposes. It was never documented and was disabled by default. As such,
we expect no one is using it and we have stopped for quite a while.

There are other tools ( https://getindico.io/ ) and services
(EventBrite) that solve this problem better than we ever did.