Fix subtle bug in StatisticsController::members()
There is a bug in the logic of the loop iterating over $targets in StatisticsController::members(), see line 320 of app/Http/Controllers/StatisticsController.php. In the case where $infra and $vlan are false, the loop will try to check whether $t->isGraphable() (line 328). But if it is not, the loop will *not* stop, and the graph $g will retain the value from the previous loop. Which means that the previous diagram $g will be pushed into $graphs and shown once more. The final effect is that, for each non-graphable member, the previous member's graph gets shown twice. This is obviously a bug. To correct, we simply add an 'else { continue; }' to go directly to the next loop.