[NF] Add support for rfc1997 pass through on route servers
RFC1997 defines some well-known communities including NO_EXPORT (0xFFFFFF01 / 65535:65281) and NO_ADVERTISE and states that they have global significance and their operations shall be implemented in any community-attribute-aware BGP speaker.

According to RFC7947, it is a matter of local policy whether these well-known communities are interpreted or passed through. Historically, some IXP route servers interpret them and some pass them through. As such the behaviour of these well-known communities is not well-understood when it comes to route servers and this topic has been the subject of a good deal of debate in the IXP community over the years.

In 2017, INEX and LONAP published draft-hilliard-grow-no-export-via-rs-00 to try and create some consensus on this. While the draft was not accepted as a standard, the discussion drew a conclusion that these well-known communities should not be interpreted by the route server but passed through.

This commit adds support to enable pass through. As enabling this on Bird resets BGP sessions, this is being added with default as off for now so as to ensure it is not a breaking / service affected change. The recommended setting is on however. We will change the default in a later version when IX's have had a chance to update their route servers in a maintenance window.

Documenation updated:

https://docs.ixpmanager.org/features/route-servers/#rfc1997-passthru