[REVIEW] user session mgmt with 2fa
In this commit, we link 2fa with session management. If a user has a 'remember me' session then they should not be asked to revalidate via 2fa when the standard session has expired but the remember me session is valid.

We needed to add an entry to the database here also to mark a remember me session as having completed 2fa and enforce it if not. Without this, it's possible to work around 2fa by deleting session cookies after login but before 2fa completed.