[IM] Fully disable standard community filtering if route server ASN is 32bit
While strictly speaking the first filter (0,peeasn) could work, the other
do will not. Better to be explicit and match the documentation and fully
disable community based filtering for route servers with a 32bit asn.

Reminder: we discourage using 32bit ASNs for route servers:

> You are strongly advised to use / request a 16-bit ASN from your RIR
> for route server use. If you do not, you will be unable to offer your
> members standard community based filtering.

Separatrely, we have a feature in development that will allow people to
set filters in the UI and which uses large BGP communities internally.