Restructure PR
Thanks for this PR @schadom - putting the MD5 on the main display is a great idea. Everythime I click through to the VLAN interface for this, I wonder why we never did it.

I've refactored the commit because I intentionally  - and naively - have IPv6 as the 'first' address in some desparate hope we'll eventually get there!

I also removed multicast - I honestly can't name an IX that supports multicast off the top of my head.