Remove references to TYPE_ROUTESERVER [release notes] - closes /inex/IXP-Manager#647
We stopped using TYPE_ROUTESERVER some time ago. It was first introduced for the IX-F member export but we subsequently added router management to IXP Manager and route servers are now identified based on there being a router with a matching IP. At the end of day, route servers are the IXP's own infrastructure and the customer/member type should reflect this.