[IM] Router server conf optimisations
NB: skinned route server configuration files will need to be updated!

This commit moves the collection of neighbour prefixes from "all on one go" to a per neighbour processing basis.

This has no immediate win but it will allow for future optimisations if/when we start streaming the router configurations to local disk for web deliver or direct to web deliver. In the current incarnation, the config is built entirely in memory which is why there is no real win with this.